Initial commit of mostly-working loader
TODO: Probably change the API of the parts to have a read_data and/or
read_strings which return the length read, and advance the buffer
pointer according to that.

If I want to promise to keep the order the same, I can check against
the buffer offsets, otherwise I need to sort them first.